refactor : 불필요한 인증 로직 제거 및 Auth0 토큰 검증 방식 단일화
