--- src/main/java/be/ddd/api/dto/req/TokenRequestDto.java
@@ -1,4 +1,0 @@
-package be.ddd.api.dto.req;
-
-public record TokenRequestDto(String refreshToken) {
-}
