--- src/main/java/be/ddd/api/dto/res/TokenResponseDto.java
@@ -1,4 +1,0 @@
-package be.ddd.api.dto.res;
-
-public record TokenResponseDto(String accessToken) {
-}
