+++ src/main/java/be/ddd/api/login/Auth0JwtVerifier.java
@@ -1,4 +1,4 @@
-package be.ddd.api.security.auth0;
+package be.ddd.api.login;
 
 import com.auth0.jwk.JwkProvider;
 import com.auth0.jwk.JwkProviderBuilder;
@@ -18,7 +18,7 @@ public class Auth0JwtVerifier {
 
     private final String issuer;
     private final String audience;
-    private final JwkProvider jwkProvider;
+    private JwkProvider jwkProvider;
 
     public Auth0JwtVerifier(
         @Value("${auth0.issuer}") String issuer,
@@ -28,9 +28,9 @@ public class Auth0JwtVerifier {
         this.issuer = issuer;
         this.audience = audience;
         this.jwkProvider = new JwkProviderBuilder(new URL("https://" + domain + "/.well-known/jwks.json"))
-                .cached(10, 24, TimeUnit.HOURS)
-                .rateLimited(10, 1, TimeUnit.SECONDS)
-                .build();
+            .cached(10, 24, TimeUnit.HOURS)
+            .rateLimited(10, 1, TimeUnit.MINUTES)
+            .build();
     }
 
     public DecodedJWT verify(String token) throws Exception {
@@ -42,7 +42,7 @@ public class Auth0JwtVerifier {
         JWTVerifier verifier = JWT.require(alg)
             .withIssuer(issuer)
             .withAudience(audience)
-            .acceptLeeway(2)
+            .acceptLeeway(3)
             .build();
 
         return verifier.verify(token);
